Refactor: файловая структура проекта
